feat(generator): 优化代码生成逻辑并添加自定义表单组件支持
- 将 List<String> 类型改为 List<Object> 类型，以支持更灵活的数据结构
- 在前端模板中添加自定义表单组件的引用和使用示例
- 优化后端控制器中的错误提示信息
- 调整数据库查询逻辑，支持按标签和值进行字典项查询

+++ continew-project/src/main/java/top/continew/admin/project/model/resp/ProjectServerConfigResp.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2022-present Charles7c Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package top.continew.admin.project.model.resp;
 
 import cn.crane4j.annotation.AssembleMethod;
@@ -9,12 +25,8 @@ import lombok.Data;
 import java.io.Serial;
 import java.util.List;
 
-import cn.crane4j.annotation.Assemble;
-import cn.crane4j.core.executor.handler.ManyToManyAssembleOperationHandler;
-
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import top.continew.admin.common.constant.ContainerConstants;
 import top.continew.admin.common.enums.DisEnableStatusEnum;
 import top.continew.admin.common.model.resp.BaseDetailResp;
 import top.continew.admin.project.service.ProjectConfigService;
@@ -65,7 +77,7 @@ public class ProjectServerConfigResp extends BaseDetailResp {
      * 服务器端口
      */
     @Schema(description = "服务器端口")
-    private Long port;
+    private Integer port;
 
     /**
      * 服务器用户名
